+++ modules/components/org.wso2.carbon.das.jobmanager.core/src/gen/java/org/wso2/carbon/das/jobmanager/core/WorkersApiService.java
@@ -1,24 +1,10 @@
-package org.wso2.carbon.analytics.jobmanager.restapi;
+package org.wso2.carbon.das.jobmanager.core;
 
-import org.wso2.carbon.analytics.jobmanager.restapi.*;
-import org.wso2.carbon.analytics.jobmanager.restapi.dto.*;
+import org.wso2.carbon.das.jobmanager.core.dto.WorkerDTO;
 
-import org.wso2.msf4j.formparam.FormDataParam;
-import org.wso2.msf4j.formparam.FileInfo;
-
-import org.wso2.carbon.analytics.jobmanager.restapi.dto.ErrorDTO;
-import org.wso2.carbon.analytics.jobmanager.restapi.dto.ExecutionPlanDTO;
-import org.wso2.carbon.analytics.jobmanager.restapi.dto.ExecutionPlanListDTO;
-import org.wso2.carbon.analytics.jobmanager.restapi.dto.WorkerDTO;
-import org.wso2.carbon.analytics.jobmanager.restapi.dto.WorkerListDTO;
-
-import java.util.List;
-import org.wso2.carbon.analytics.jobmanager.restapi.NotFoundException;
-
-import java.io.InputStream;
+import org.wso2.carbon.das.jobmanager.core.dto.ExecutionPlanDTO;
 
 import javax.ws.rs.core.Response;
-import javax.ws.rs.core.SecurityContext;
 
 @javax.annotation.Generated(value = "org.wso2.maven.plugins.JavaMSF4JServerCodegen", date = "2017-03-14T10:36:41.439+05:30")
 public abstract class WorkersApiService {
