+++ modules/components/org.wso2.carbon.das.jobmanager.core/src/gen/java/org/wso2/carbon/das/jobmanager/core/dto/ErrorDTO.java
@@ -1,13 +1,10 @@
-package org.wso2.carbon.analytics.jobmanager.restapi.dto;
+package org.wso2.carbon.das.jobmanager.core.dto;
 
 import java.util.Objects;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import java.util.ArrayList;
 import java.util.List;
-import org.wso2.carbon.analytics.jobmanager.restapi.dto.ErrorListItemDTO;
 
 /**
  * ErrorDTO
