+++ modules/components/org.wso2.carbon.das.jobmanager.core/src/gen/java/org/wso2/carbon/das/jobmanager/core/dto/WorkerDTO.java
@@ -1,9 +1,7 @@
-package org.wso2.carbon.analytics.jobmanager.restapi.dto;
+package org.wso2.carbon.das.jobmanager.core.dto;
 
 import java.util.Objects;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 
 /**
