+++ modules/components/org.wso2.carbon.das.jobmanager.core/src/main/java/org/wso2/carbon/das/jobmanager/core/factories/WorkersApiServiceFactory.java
@@ -1,0 +1,12 @@
+package org.wso2.carbon.das.jobmanager.core.factories;
+
+import org.wso2.carbon.das.jobmanager.core.WorkersApiService;
+import org.wso2.carbon.das.jobmanager.core.impl.WorkersApiServiceImpl;
+
+public class WorkersApiServiceFactory {
+    private static final WorkersApiService service = new WorkersApiServiceImpl();
+
+    public static WorkersApiService getWorkersApi() {
+        return service;
+    }
+}
